Major change before final release
- Fixed default sound and sound type in Alarm API
- Finished and finalised welcome screen
- Finished recipient item
- Finished add bookmark for Safari

+++ API/AlarmSubstrate/interface.h
@@ -36,7 +36,7 @@ typedef enum {
 
 @interface AlarmManager : NSObject
 
-@property(readonly, nonatomic) int defaultSoundType;
+@property(readonly, nonatomic) NSInteger defaultSoundType;
 @property(readonly, nonatomic) NSString *defaultSound;
 
 + (NSArray *)copyReadAlarmsFromPreferences;
